requirements.txt
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
aiohttp>=3.9.0
pyarrow>=14.0.0
pyyaml>=6.0.0
python-dateutil>=2.8.0
